Show a dialog requiring restart after enabling/disabling extension
When the extension is enabled for the first time, set a restart flag that is resetted by the looper_interface when VLC is restarted. This allows to show a dialog to invite the user to restart every time he tries to enable the extension.

If the extension is disabled instead, the user can reenable it and will see the restart dialog again.