Add check to avoid restart dialog loop
If VLC crashes when the extension is enabled the first time, it is possible that the restart flag is set to false but the interface is not activated, so that it will never set the restart flag to true.

At this point every time the user tries to activate the extension the restart dialog will appear and the only solution is to edit the configuration manually.